INV-4: Add simple component line  - Modify: rename lineClient on lineFetch.ts  - Modify: rename import on LineComponent.ts  - Modify: extract interface on LineAPI.ts  - Modify: rename lineclient on lineClient.ts
