+++ src/business/line/api/lineClient.ts
@@ -1,5 +1,5 @@
 import axios from 'axios';
 
-export const client = axios.create({
+export const lineClient = axios.create({
     baseURL: "http://91.108.126.99:8080/v1/lines"
 })
